requirements.txt
asttokens==2.2.0
backcall==0.2.0
colorama==0.4.6
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
numpy-stl==2.17.1
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
plotly==5.11.0
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.6
pywin32==305
pyzmq==24.0.1
six==1.16.0
stack-data==0.6.2
tenacity==8.1.0
tornado==6.2
traitlets==5.6.0
wcwidth==0.2.5
